+++ R/zzz.R
@@ -30,8 +30,9 @@
 #       together with declaring the required packaged in the "Imports" section of the DESCRIPTION file.
 #       Source: https://cran.r-project.org/doc/manuals/R-exts.html#Specifying-imports-and-exports
 
-#' @importFrom futile.logger flog.error flog.warn flog.info
-#' @importFrom utils         dump.frames
+# Imports DISABLED since there is longer an "Imports" dependency on futile.logger (since version 1.1.0):
+# @importFrom futile.logger flog.error flog.warn flog.info
+# @importFrom utils         dump.frames
 # NULL
 
 
@@ -59,6 +60,17 @@
   # packageStartupMessage(paste("Library path (libname):", libname))
 
   .tryCatchLog.env$newline <- determine.platform.NewLine()
+
+
+  
+  # Decide which logging functions to use
+  if (is.package.available("futile.logger")) {
+    packageStartupMessage("Using futile.logger for logging...")
+    set.logging.functions(futile.logger::flog.error, futile.logger::flog.warn, futile.logger::flog.info)
+  } else {
+    packageStartupMessage("futile.logger not found. Using tryCatchLog-internal functions for logging...")
+    set.logging.functions()    # Default: Activate the package-internal minimal logging functions
+  }
 
 }