+++ tests/testthat/test_limited_Labels_Compact.R
@@ -1,10 +1,11 @@
-# library(futile.logger)
 # library(tryCatchLog)
-# library(testthat)
+library(testthat)
 
 
 
-context("limitedLabelsCompact")
+context("test_limited_Labels_Compact.R")
+
+
 
 # set up test context
 options("tryCatchLog.write.error.dump.file" = FALSE)
@@ -13,8 +14,8 @@ options("tryCatchLog.silent.messages"     = FALSE)
 
 
 
-# flog.threshold("FATAL")                         # suppress logging of errors and warnings to avoid overly output
-# flog.threshold("INFO")
+# suppress logging of errors and warnings to avoid overly output
+# source("disable_logging_output.R")