+++ tests/testthat/test_silent_messages.R
@@ -1,13 +1,15 @@
 # Tests are run within the folder "tryCatchLog/tests/testthat".
 # Clean it up at the beginning of a test!
 
-library(futile.logger)
+library(testthat)
 
 
 
 # Silent warnings -------------------------------------------------------------------------------------------------
 
-context("Silent messages")
+context("test_silent_messages.R")
+
+
 
 # set up test context
 options("tryCatchLog.write.error.dump.file" = FALSE)
@@ -16,7 +18,8 @@ options("tryCatchLog.silent.messages"     = FALSE)
 
 
 
-flog.threshold("FATAL")                         # suppress logging of errors and warnings to avoid overly output
+# suppress logging of errors and warnings to avoid overly output
+source("disable_logging_output.R")