+++ tests/testthat/test_silent_warnings.R
@@ -1,13 +1,15 @@
 # Tests are run within the folder "tryCatchLog/tests/testthat".
 # Clean it up at the beginning of a test!
 
-library(futile.logger)
+library(testthat)
 
 
 
 # Silent warnings -------------------------------------------------------------------------------------------------
 
-context("Silent warnings")
+context("test_silent_warnings.R")
+
+
 
 # set up test context
 options("tryCatchLog.write.error.dump.file" = FALSE)    # global default setting for all tryCatchLog call params "write.error.dump.file"
@@ -15,7 +17,9 @@ options("tryCatchLog.silent.warnings" = FALSE)
 options("tryCatchLog.silent.messages" = FALSE)
 
 
-flog.threshold("FATAL")                               # suppress logging of errors and warnings to avoid overly output
+
+# suppress logging of errors and warnings to avoid overly output
+source("disable_logging_output.R")