+++ tests/testthat/test_tryCatch_semantics.R
@@ -3,13 +3,12 @@
 
 library(testthat)
 library(tryCatchLog)
-library(futile.logger)
 
 
 
 # inits -----------------------------------------------------------------------------------------------------------
 
-context("tryCatch semantics")
+context("test_tryCatch_semantics.R")
 
 
 
@@ -26,7 +25,8 @@ options("tryCatchLog.silent.messages" = FALSE)
 
 
 
-flog.threshold("FATAL")                               # suppress logging of errors and warnings to avoid overly output
+# suppress logging of errors and warnings to avoid overly output
+source("disable_logging_output.R")