+++ tests/testthat/test_tryLog_basics.R
@@ -1,18 +1,13 @@
 library(tryCatchLog)
 library(testthat)
-library(futile.logger)
-
-# Basic tests of tryLog
-
-# Tests are run within the folder "tryCatchLog/tests/testthat".
-# Clean it up at the beginning of a test!
-
 
 
 
 # Basic tests -----------------------------------------------------------------------------------------------------
 
-context("tryLog")
+context("test_tryLog_basics.R")
+
+
 
 # set up test context
 options("tryCatchLog.write.error.dump.file" = FALSE)    # global default setting for all tryCatchLog call params "write.error.dump.file"
@@ -21,7 +16,8 @@ options("tryCatchLog.silent.messages" = FALSE)
 
 
 
-flog.threshold("FATAL")                               # suppress logging of errors and warnings to avoid overly output
+# suppress logging of errors and warnings to avoid overly output
+source("disable_logging_output.R")